evals/requirements.txt
anthropic>=0.45.0
braintrust>=0.0.180
httpx>=0.27.0
python-dotenv>=1.0.0
